refactor: move config from configs to websocket
